Finished auth flow, redesign of email invite template and UI
