+++ src/app/admin/(protected)/reports/page.js
@@ -4,7 +4,7 @@ import OperationsReport from "@/components/OperationsReport";
 export const dynamic = "force-dynamic";
 
 export const metadata = {
-  title: "Admin Reports | Walhez",
+  title: "Admin Reports",
   description: "Protected equipment income and expense reporting for administrators.",
 };
