+++ src/app/operations-report/page.js
@@ -2,7 +2,7 @@ import { redirect } from "next/navigation";
 import { requireAdmin } from "@/lib/auth";
 
 export const metadata = {
-  title: "Operations Report | Walhez",
+  title: "Operations Report",
   description: "Redirects to the protected admin report area.",
 };
