+++ src/app/operations-report/pdf/page.js
@@ -2,7 +2,7 @@ import { redirect } from "next/navigation";
 import { requireAdmin } from "@/lib/auth";
 
 export const metadata = {
-  title: "Operations Report PDF | Walhez",
+  title: "Operations Report PDF",
   description: "Redirects to the protected admin PDF report area.",
 };
