+++ src/app/services/page.js
@@ -1,5 +1,10 @@
 import React from 'react'
 
+export const metadata = {
+  title: "Services",
+  description: "Browse Walhez Group service offerings across construction support and equipment operations.",
+};
+
 export default function page(){
   return(
     <div>
